refactor: add few minor changes
