--- src/models/game-play-models.ts
@@ -1,6 +1,0 @@
-export interface AttackRequest {
-  gameId: number;
-  x: number;
-  y: number;
-  indexPlayer: number;
-}
