+++ src/models/player-models.ts
@@ -81,3 +81,10 @@ export interface Win {
   name: string;
   wins: number;
 }
+
+export interface AttackRequest {
+  gameId: number;
+  x: number;
+  y: number;
+  indexPlayer: number;
+}
